content/media/video/test/test_progress2.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: progress events</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
// Test progress events with wav backend
var completed = false;
var last_progress = false;

function on_loadedmetadata() {
  var v = document.getElementById('v');
  v.play();
}

function do_progress(e) {
  ok(!completed, "Check for progress event after completed: " + completed);
  ok(e.lengthComputable, "Check progress lengthComputable");
  ok(e.loaded >= 0 && e.loaded <= e.total, "Check progress loaded: " + e.loaded);
  ok(e.total == 102444, "Check progress total: " + e.total);
  last_progress = e.loaded;
}

function do_ended() {
  ok(!completed, "Check for duplicate ended event");
  completed = true;
  ok(last_progress == 102444, "Last progress event size: " + last_progress);
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<audio id='v'
       src='big.wav'
       onloadedmetadata='on_loadedmetadata()'
       onended='do_ended()'
       onprogress='do_progress(event)'>
</audio>
</body>
</html>
